Final changes

+++ main.py
@@ -1,4 +1,4 @@
-from config import TOKEN
+from constants import TOKEN
 import logging, controller, checks, constants
 
 
@@ -15,9 +15,7 @@ from telegram.ext import (
 )
 
 
-logging.basicConfig(
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO
-)
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO, filename= constants.LOG_NAME, encoding='utf-8')
 logger = logging.getLogger(__name__)
 
 
@@ -63,7 +61,7 @@ def Name(update, _):
     return -> int (переход на следующий шаг по списку)
     """
     user = update.message.from_user
-    constants.NAME = update.message.text
+    constants.CONTACT_DATA['Имя'] = update.message.text
     if checks.Check_name(update) == False: return STEP_NAME
     logger.info("Пользователь %s добавил имя: %s", user.first_name, update.message.text)
     update.message.reply_text('Введите фамилию.\n' 'Команда /cancel, чтобы прекратить разговор.')
@@ -76,7 +74,7 @@ def Last_name(update, _):
     return -> int (переход на следующий шаг по списку)
     """
     user = update.message.from_user
-    constants.LAST_NAME = update.message.text
+    constants.CONTACT_DATA['Фамилия'] = update.message.text
     if checks.Check_name(update) == False: return STEP_LAST_NAME
     logger.info("Пользователь %s добавил фамилию: %s", user.first_name, update.message.text)
     update.message.reply_text('Введите номер телефона.\n' 'Команда /cancel, чтобы прекратить разговор.')
@@ -90,7 +88,7 @@ def Phone_num(update, _):
     user = update.message.from_user
     if checks.Check_num(update) == False: return STEP_PHONE_NUM
     if checks.Check_phone(update) == False: return STEP_PHONE_NUM
-    constants.PHONE_NUM = update.message.text
+    constants.CONTACT_DATA['Телефон'] = update.message.text
     logger.info("Пользователь %s добавил номер телефона: %s", user.first_name, update.message.text)
     update.message.reply_text('Введите группу или отправь /skip, если без группы.\n' 'Команда /cancel, чтобы прекратить разговор.')
     return STEP_GROUP
@@ -101,7 +99,7 @@ def Group(update, _):
     return -> int (переход на следующий шаг по списку)
     """
     user = update.message.from_user
-    constants.GROUP = update.message.text
+    constants.CONTACT_DATA['Группа'] = update.message.text
     logger.info("Пользователь %s добавил группу: %s", user.first_name, update.message.text)
     update.message.reply_text('Введите комментарий или отправь /skip, если без комментария.\n' 'Команда /cancel, чтобы прекратить разговор.')
     return STEP_COMMENT
@@ -122,7 +120,7 @@ def Comment(update, _):
     return -> завершение ConversationHandler
     """
     user = update.message.from_user
-    constants.COMMENT = update.message.text
+    constants.CONTACT_DATA['Комментарий'] = update.message.text
     logger.info("Пользователь %s добавил Комментарий: %s", user.first_name, update.message.text)
     update.message.reply_text('Данные сохранены!\n' 'Для перехода в меню отправьте /start')
     controller.add_contact()
@@ -146,7 +144,7 @@ def Search(update, _):
     """
     user = update.message.from_user
     if checks.Check_name(update) == False: return SEARCH
-    # if checks.Check_none(update) == False: return SEARCH
+    if checks.Check_none(update) == True: return SEARCH
     logger.info("Пользователь %s искал: %s", user.first_name, update.message.text)
     update.message.reply_text(f'{controller.Search_contact(update)} \n''Для перехода в меню отправьте /start')
     return ConversationHandler.END 
@@ -159,7 +157,7 @@ def Delete(update, _):
     user = update.message.from_user
     logger.info("Пользователь %s удалил контакт: %s", user.first_name, update.message.text)
     if checks.Check_name(update) == False: return DELETE
-    # if checks.Check_none(update) == False: return DELETE
+    if checks.Check_none(update) == True: return DELETE
     update.message.reply_text(f'{controller.Search_contact(update)} \n''Введите номер телефона контакта для удаления\n')
     return DELETE_SECOND
    
@@ -170,7 +168,7 @@ def Delete_second(update, _):
     return -> завершение ConversationHandler
     """
     if checks.Check_num(update) == False: return DELETE_SECOND
-    # if checks.Check_phone_del(update) == False: return DELETE_SECOND
+    if checks.Check_phone_del(update) == True: return DELETE_SECOND
     controller.Delete_contact(update)
     update.message.reply_text('Контакт удален \n''Для перехода в меню отправьте /start')
     return ConversationHandler.END
